Adding the second master.
With the addition of a second master to the file, the initial base letters have been created for the regular weight. They include the group of vowels and consonants of the Bangla script. Also, the name of the file has been changed from "Alkatra-Bold" to "Alkatra-Bangla", due to the addition of multiple masters.